Support stack size system property
The Java stack size may be specified using a system property. Use
a specific system property for stack size rather than opening up all
Java command line options to be specifiable.

There is no validation of the input stack size except to ensure that
it does not contain whitespace. The input stack size, if valid, is
appended to -Xss= and added to the java command line returned from
release.

Add stack size documentation to the README.

[#49694449]